+++ src/BOOST/MATH/cmake/Dependencies.cmake
@@ -1,0 +1,6 @@
+botgPackageDependencies(
+    LIB_REQUIRED_TPLS
+        BOOST_MATH
+    TEST_REQUIRED_PACKAGES
+        BootsOnTheGround_GTEST
+)